Move mempool entry & fee policy estimator into separate file
