+++ divi/src/Makefile.am
@@ -313,6 +313,8 @@ BITCOIN_CORE_H = \
   tinyformat.h \
   torcontrol.h \
   txdb.h \
+  MemPoolEntry.h \
+  FeePolicyEstimator.h \
   txmempool.h \
   ui_interface.h \
   uint256.h \
@@ -467,6 +469,8 @@ libbitcoin_server_a_SOURCES = \
   timedata.cpp \
   torcontrol.cpp \
   txdb.cpp \
+  MemPoolEntry.cpp \
+  FeePolicyEstimator.cpp \
   txmempool.cpp \
   NotificationInterface.cpp \
   version.cpp \
