+++ divi/src/txmempool.h
@@ -16,6 +16,7 @@
 #include "coins.h"
 #include "primitives/transaction.h"
 #include "sync.h"
+#include <MemPoolEntry.h>
 
 class BlockMap;
 class CAutoFile;
@@ -36,35 +37,7 @@ inline bool AllowFree(double dPriority)
 /** Fake height value used in CCoins to signify they are only in the memory pool (since 0.8) */
 bool IsMemPoolHeight(unsigned coinHeight);
 
-/**
- * CTxMemPool stores these:
- */
-class CTxMemPoolEntry
-{
-private:
-    CTransaction tx;
-    CAmount nFee;         //! Cached to avoid expensive parent-transaction lookups
-    size_t nTxSize;       //! ... and avoid recomputing tx size
-    size_t nModSize;      //! ... and modified size for priority
-    int64_t nTime;        //! Local time when entering the mempool
-    double dPriority;     //! Priority when entering the mempool
-    unsigned int nHeight; //! Chain height when entering the mempool
-
-public:
-    CTxMemPoolEntry(const CTransaction& _tx, const CAmount& _nFee, int64_t _nTime, double _dPriority, unsigned int _nHeight);
-    CTxMemPoolEntry();
-    CTxMemPoolEntry(const CTxMemPoolEntry& other);
-
-    const CTransaction& GetTx() const { return this->tx; }
-    double GetPriority(unsigned int currentHeight) const;
-    CAmount GetFee() const { return nFee; }
-    size_t GetTxSize() const { return nTxSize; }
-    int64_t GetTime() const { return nTime; }
-    unsigned int GetHeight() const { return nHeight; }
-};
-
 class CMinerPolicyEstimator;
-
 /** An inpoint - a combination of a transaction and an index n into its vin */
 class CInPoint
 {
